[REFACTOR] Moved sign up form
